pull up AddDefaultTopN
