+++ x-pack/plugin/esql/src/main/java/org/elasticsearch/xpack/esql/optimizer/LogicalPlanOptimizer.java
@@ -59,6 +59,7 @@ import org.elasticsearch.xpack.esql.expression.function.scalar.nulls.Coalesce;
 import org.elasticsearch.xpack.esql.expression.function.scalar.spatial.SpatialRelatesFunction;
 import org.elasticsearch.xpack.esql.expression.predicate.operator.comparison.Equals;
 import org.elasticsearch.xpack.esql.expression.predicate.operator.comparison.In;
+import org.elasticsearch.xpack.esql.optimizer.rules.AddDefaultTopN;
 import org.elasticsearch.xpack.esql.optimizer.rules.BooleanFunctionEqualsElimination;
 import org.elasticsearch.xpack.esql.optimizer.rules.CombineDisjunctionsToIn;
 import org.elasticsearch.xpack.esql.optimizer.rules.ConstantFolding;
@@ -1286,40 +1287,6 @@ public class LogicalPlanOptimizer extends ParameterizedRuleExecutor<LogicalPlan,
         }
     }
 
-    /**
-     * This adds an explicit TopN node to a plan that only has an OrderBy right before Lucene.
-     * To date, the only known use case that "needs" this is a query of the form
-     * from test
-     * | sort emp_no
-     * | mv_expand first_name
-     * | rename first_name AS x
-     * | where x LIKE "*a*"
-     * | limit 15
-     *
-     * or
-     *
-     * from test
-     * | sort emp_no
-     * | mv_expand first_name
-     * | sort first_name
-     * | limit 15
-     *
-     * PushDownAndCombineLimits rule will copy the "limit 15" after "sort emp_no" if there is no filter on the expanded values
-     * OR if there is no sort between "limit" and "mv_expand".
-     * But, since this type of query has such a filter, the "sort emp_no" will have no limit when it reaches the current rule.
-     */
-    static class AddDefaultTopN extends ParameterizedOptimizerRule<LogicalPlan, LogicalOptimizerContext> {
-
-        @Override
-        protected LogicalPlan rule(LogicalPlan plan, LogicalOptimizerContext context) {
-            if (plan instanceof UnaryPlan unary && unary.child() instanceof OrderBy order && order.child() instanceof EsRelation relation) {
-                var limit = new Literal(plan.source(), context.configuration().resultTruncationMaxSize(), DataType.INTEGER);
-                return unary.replaceChild(new TopN(plan.source(), relation, order.order(), limit));
-            }
-            return plan;
-        }
-    }
-
     public static class ReplaceRegexMatch extends org.elasticsearch.xpack.esql.core.optimizer.OptimizerRules.OptimizerExpressionRule<
         RegexMatch<?>> {
 
@@ -1735,7 +1702,7 @@ public class LogicalPlanOptimizer extends ParameterizedRuleExecutor<LogicalPlan,
         }
     }
 
-    private abstract static class ParameterizedOptimizerRule<SubPlan extends LogicalPlan, P> extends ParameterizedRule<
+    public abstract static class ParameterizedOptimizerRule<SubPlan extends LogicalPlan, P> extends ParameterizedRule<
         SubPlan,
         LogicalPlan,
         P> {
